feat(xs): #616 give 靈感 and 我的 separate URL paths (/ and /my)
