+++ apps/product/src/app/[locale]/(with-layout)/my/page.tsx
@@ -1,0 +1,189 @@
+"use client";
+
+import { useMyPracticeStats, useMyPractices } from "@daodao/api";
+import { MessagesSvg } from "@daodao/assets";
+import { useRouter } from "@daodao/i18n/navigation";
+import { cn } from "@daodao/ui/lib/utils";
+import { CheckCircle2 } from "lucide-react";
+import { useMemo, useState } from "react";
+import {
+  AddTaskFAB,
+  DashboardHeader,
+  InProgressSection,
+  type InProgressTask,
+  RecommendationSection,
+} from "@/components/dashboard";
+import { BackgroundAnimation, Banner } from "@/components/layout";
+import { RandomPracticesSection } from "@/components/practice/shared/random-practices-section";
+import {
+  FilterStatus,
+  type FilterStatus as FilterStatusType,
+  mapPracticeStatusToTaskStatus,
+} from "@/constants/task-status";
+import { HOME_TAB_PATHS } from "@/constants/home-navigation";
+
+const filterOptions = [
+  { value: FilterStatus.all, label: "全部" },
+  { value: FilterStatus.draft, label: "草稿" },
+  { value: FilterStatus.notStarted, label: "未開始" },
+  { value: FilterStatus.inProgress, label: "進行中" },
+  { value: FilterStatus.completed, label: "已完成" },
+];
+
+export default function MyPage() {
+  const router = useRouter();
+  const [filterStatus, setFilterStatus] = useState<FilterStatusType>(FilterStatus.all);
+
+  const { data: allPracticesData, isLoading: isMyLoading } = useMyPractices({ limit: 16 });
+  const { data: statsData } = useMyPracticeStats();
+
+  const { inProgressTasks } = useMemo(() => {
+    const practices = allPracticesData?.data || [];
+    const inProgressTasksData: InProgressTask[] = [];
+
+    practices.forEach((practice) => {
+      const lastCheckInDate = practice.lastCheckinAt ?? null;
+      inProgressTasksData.push({
+        id: practice.id,
+        label: "主題實踐",
+        title: practice.title,
+        description: practice.practiceAction || "",
+        checkInCount: practice.checkInCount,
+        progress: practice.progressPercentage ?? 0,
+        messagesCount: 0,
+        isUnreadMessages: false,
+        theme: practice.themeColor || "#FCDD84",
+        status: mapPracticeStatusToTaskStatus(practice.status),
+        lastCheckInDate,
+        startDate: practice.startDate || null,
+        endDate: practice.endDate || null,
+      });
+    });
+
+    return { inProgressTasks: inProgressTasksData };
+  }, [allPracticesData]);
+
+  const filteredInProgressTasks = useMemo(() => {
+    if (filterStatus === FilterStatus.completed)
+      return inProgressTasks.filter((task) => task.status === FilterStatus.completed);
+    if (filterStatus === FilterStatus.all)
+      return inProgressTasks.filter((task) => task.status !== FilterStatus.completed);
+    return inProgressTasks.filter((task) => task.status === filterStatus);
+  }, [inProgressTasks, filterStatus]);
+
+  const stats = useMemo(() => {
+    const statsDataValue = statsData?.data;
+    return [
+      {
+        label: "連續登入",
+        value: String(statsDataValue?.currentStreak || 0),
+        unit: "天",
+        icon: CheckCircle2,
+      },
+      {
+        label: "獲得迴響",
+        value: String(statsDataValue?.totalCheckIns || 0),
+        unit: "次",
+        icon: MessagesSvg,
+      },
+    ];
+  }, [statsData]);
+
+  const hasPractices = inProgressTasks.length > 0;
+
+  const filterCounts = useMemo(() => {
+    const counts = {
+      [FilterStatus.all]: 0,
+      [FilterStatus.draft]: 0,
+      [FilterStatus.notStarted]: 0,
+      [FilterStatus.inProgress]: 0,
+      [FilterStatus.completed]: 0,
+    };
+    for (const task of inProgressTasks) {
+      if (task.status !== FilterStatus.completed) {
+        counts[FilterStatus.all]++;
+      }
+      if (task.status in counts) {
+        counts[task.status as keyof typeof counts]++;
+      }
+    }
+    return counts;
+  }, [inProgressTasks]);
+
+  return (
+    <div className="relative min-h-screen">
+      <Banner />
+      <BackgroundAnimation />
+
+      <main className="relative z-[25] pb-[72px] bg-very-light-gray">
+        <div className="max-w-[640px] px-4 mx-auto pt-4">
+          {/* Tab Switcher */}
+          <div className="flex border-b border-[#E5E7EB] mb-4">
+            <button
+              type="button"
+              onClick={() => router.push(HOME_TAB_PATHS.inspire)}
+              className={cn(
+                "flex-1 py-2 text-sm font-medium transition-all",
+                "text-text-dark/40"
+              )}
+            >
+              靈感
+            </button>
+            <button
+              type="button"
+              className={cn(
+                "flex-1 py-2 text-sm font-medium transition-all",
+                "text-text-dark border-b-2 border-logo-cyan -mb-px"
+              )}
+            >
+              我的
+            </button>
+          </div>
+
+          {isMyLoading ? (
+            <div className="text-center text-text-dark">載入中...</div>
+          ) : (
+            <>
+              <DashboardHeader stats={stats} />
+              {!hasPractices && <RandomPracticesSection compact />}
+              {hasPractices && (
+                <>
+                  <div className="mb-4">
+                    <div
+                      role="tablist"
+                      aria-label="任務篩選"
+                      className="flex gap-2 overflow-x-auto scrollbar-hide"
+                    >
+                      {filterOptions.map((option) => (
+                        <button
+                          type="button"
+                          key={option.value}
+                          role="tab"
+                          aria-selected={filterStatus === option.value}
+                          onClick={() => setFilterStatus(option.value)}
+                          className={cn(
+                            "px-5 py-2 rounded-full text-sm whitespace-nowrap border transition-colors",
+                            filterStatus === option.value
+                              ? "bg-primary-base border-primary-base text-white"
+                              : "bg-white border-primary-base text-primary-base"
+                          )}
+                        >
+                          {option.label} {filterCounts[option.value]}
+                        </button>
+                      ))}
+                    </div>
+                  </div>
+
+                  <InProgressSection tasks={filteredInProgressTasks} />
+                  <RecommendationSection onGoToInspire={() => router.push(HOME_TAB_PATHS.inspire)} />
+                </>
+              )}
+            </>
+          )}
+        </div>
+      </main>
+
+      <AddTaskFAB onAddTask={() => router.push("/practices/create")} />
+    </div>
+  );
+}
