+++ apps/product/src/app/[locale]/practices/[id]/page.tsx
@@ -28,6 +28,7 @@ import type { IComment, ICommentReply } from "@/components/check-in/reactions";
 import { BackgroundAnimation } from "@/components/layout";
 import { PracticeDetailShell } from "@/components/practice";
 import { applyOnboardingUpdateFromResponse } from "@/components/task-guide/onboarding-progress-context";
+import { HOME_TAB_PATHS } from "@/constants/home-navigation";
 import {
   type DurationDays,
   DurationDays as DurationDaysConst,
@@ -469,7 +470,7 @@ export default function PracticeDetailPage() {
         <div className="sticky top-0 z-50 max-w-[448px] mx-auto w-full">
           <button
             type="button"
-            onClick={() => router.replace("/?tab=mine")}
+            onClick={() => router.replace(HOME_TAB_PATHS.mine)}
             className="absolute top-2 right-2 flex items-center justify-center size-10 rounded-full text-light-gray bg-very-light-gray/50 hover:text-logo-cyan"
             aria-label="關閉"
           >
@@ -490,7 +491,7 @@ export default function PracticeDetailPage() {
         <div className="sticky top-0 z-50 max-w-[448px] mx-auto w-full">
           <button
             type="button"
-            onClick={() => router.replace("/?tab=mine")}
+            onClick={() => router.replace(HOME_TAB_PATHS.mine)}
             className="absolute top-2 right-2 flex items-center justify-center size-10 rounded-full text-light-gray bg-very-light-gray/50 hover:text-logo-cyan"
             aria-label="關閉"
           >
@@ -520,7 +521,7 @@ export default function PracticeDetailPage() {
       <div className="sticky top-0 z-50 max-w-[448px] mx-auto w-full">
         <button
           type="button"
-          onClick={() => router.replace("/?tab=mine")}
+          onClick={() => router.replace(HOME_TAB_PATHS.mine)}
           className="absolute top-2 right-2 flex items-center justify-center size-10 rounded-full text-light-gray bg-very-light-gray/50 hover:text-logo-cyan"
           aria-label="關閉"
         >
