+++ apps/product/src/constants/home-navigation.ts
@@ -1,0 +1,4 @@
+export const HOME_TAB_PATHS = {
+  inspire: "/",
+  mine: "/my",
+} as const;
